Make a separate function for everything cuz of the way JS works.
JavaScript Loop Only Uses the Last Value cuz When you use a variable that is declared outside the scope the variable is going to be used in, it will use the value that variable has at the time it runs. It doesn't get a copy of the value at the time the closure is setup.

Signed-off-by: Sid-Sun <sid-sun@outlook.com>